+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGE LOG
 
-## [Pre-release]
+## [0.17.1]
 - Replace dgx-header-component and DS Footer with Header/Footer components from nypl-header-app 
 - Refactored Playwright/Cucumber tests to use new click action function